move redis to store
