+++ packages/sockethub/src/common.ts
@@ -1,23 +1,5 @@
-import SecureStore from 'secure-store-redis';
-
-import config from "./config";
 import crypto from "./crypto";
-import {JobDataDecrypted, JobEncrypted} from "./sockethub";
-
-export interface ISecureStoreInstance {
-  save(id: string, obj: any, cb: Function);
-  get(id: string, cb: Function);
-}
-
-export function getSessionStore(parentId: string, parentSecret: string,
-                                sessionId: string, sessionSecret: string,
-                                StoreObject: SecureStore = SecureStore): ISecureStoreInstance {
-  return new StoreObject({
-    namespace: 'sockethub:' + parentId + ':session:' + sessionId + ':store',
-    secret: parentSecret + sessionSecret,
-    redis: config.get('redis')
-  });
-}
+import { JobDataDecrypted, JobEncrypted } from "./sockethub";
 
 export function getPlatformId(platform: string, actor?: string): string {
   return actor ? crypto.hash(platform + actor) : crypto.hash(platform);
